+++ packages/webapp/src/inductions/components/induction-journeys/common/index.ts
@@ -1,5 +1,6 @@
 export * from "./endorsements-status";
 export * from "./induction-expires-in";
+export * from "./induction-names";
 export * from "./induction-step";
 export * from "./induction-steps-container";
 export * from "./member-card-preview";
